remove/ui/style (#23)
* style: Remove unnecessary styles from float button

* refactor(ui): Change div to span in tag rendering

Changed the tag rendering from using a div element to a span element for better
semantic meaning. This change improves the structure and accessibility of the
component.

* style: Remove unnecessary styles from float button
refactor(ui): Change div to span in tag rendering

Changed the tag rendering from using a div element to a span element for better
semantic meaning. This change improves the structure and accessibility of the
component.

* style: Remove unnecessary styles from float button
refactor(ui): Change div to span in tag rendering

Changed the tag rendering from using a div element to a span element for better
semantic meaning. This change improves the structure and accessibility of the
component.
Merged origin/remove/ui/style into remove/ui/style